feat(embedded-studio): enable manifest extraction
