+++ dev/embedded-studio/sanity.cli.ts
@@ -1,0 +1,9 @@
+import {defineCliConfig} from 'sanity/cli'
+
+export default defineCliConfig({
+  api: {
+    projectId: 'ppsg7ml5',
+    dataset: 'test',
+  },
+  unstable_staticAssetsPath: './dist/assets',
+})
